requirements.txt
# Django Core
Django>=4.2.0
django-ninja>=1.0.0
django-cors-headers>=4.3.0
gunicorn>=21.0.0

# PostgreSQL
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# WebSocket Support
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# Celery for async tasks
celery>=5.3.0
redis>=5.0.0

# ML Models - Speech Recognition & Diarization
openai-whisper==20231117
# Pin to specific versions for compatibility
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
# pyannote.audio 3.1.1 is compatible with torch 2.1.0 and numpy 1.x
pyannote.audio==3.1.1
numpy<2.0  # pyannote.audio 3.1.1 requires numpy 1.x
# Pin huggingface-hub to last version supporting use_auth_token (pyannote.audio 3.1.1 compatibility)
huggingface-hub<1.0.0

# Audio Processing
pydub>=0.25.1
soundfile>=0.12.1
librosa>=0.10.0

# Utilities
python-dotenv>=1.0.0
